Basic support for postgres schemas
All table exports that are not in the public schema are prefixed with
the schema name. This is the way Kysely is designed to work an
"enumerable" amount of schemas.

The table/model naming logic is moved to the dialect since schemas
in other dialects (like mysql) are very different and should be
treated differently. Schemas in mysql are the same as databases
in postgres.