feat: support libsql/sqld
libsql is a fork of SQLite used by Turso.tech and others.
Kysely dialect for it was contributed by https://github.com/libsql/kysely-libsql
and this commit adds support for it kysely-codegen.

The alternative to leverage codegen is to pull the schema of your libsql
database, put it into a sqlite file and then run the codegen.